- Add support for Dynamic Frameworks and Carthage - Convert BVSDKDemo to Swift 3 syntax - Cleanup compiler warnings - Remove deprecated Conversations API from BVSDK 2.x - Update Gimbal.framework to 2.52.1 - Fix use of %lu and %li in some spots - Fix missing negation operator when checking for array type in the BVRatingDistribution - Serveral updates for ci scripts and matrixed travis ci builds

+++ Pod/BVConversations/Submission/Answer/BVAnswerSubmission.m
@@ -151,7 +151,7 @@
             NSDictionary *json = [NSJSONSerialization JSONObjectWithData:data options:kNilOptions error:&jsonParsingError];
             BVAnswerSubmissionErrorResponse* errorResponse = [[BVAnswerSubmissionErrorResponse alloc] initWithApiResponse:json]; // fails gracefully
         
-            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%d)", json, statusCode]];
+            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%ld)", json, (long)statusCode]];
             
             if (httpError) {
                 // network error was generated
